refractor quick scan
